+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.4
 
 - Support stable `text` control nestable under `area`
+- Fix `color` property support for `string` to accept 255-based rgb values
 - Fix issue with alternating string colors in examples/basic_draw_text.rb
 
 ## 0.2.3